Configs for PrintToString was added
