+++ ObjectPrinting/MemberExtensions.cs
@@ -16,17 +16,3 @@ public static class MemberExtensions
         return propertyConfig.Configure(cultureInfo);
     }
 }
-
-public static class TypeExtensions
-{
-    public static IBasicConfigurator<TOwner> TrimByLength<TOwner>(this ITypeConfigurator<TOwner, string> typeConfig, int length)
-    {
-        return typeConfig.Configure(length);
-    }
-
-    public static IBasicConfigurator<TOwner> SetCulture<TOwner>(this ITypeConfigurator<TOwner, int> typeConfig,
-        CultureInfo cultureInfo)
-    {
-        return typeConfig.Configure(cultureInfo);
-    }
-}
